#657474: Deleting a custom task designated as the site home page no longer deletes the site_frontpage variable.
Deleting the site_frontpage variable alongside the custom task that had previously owned it worked OK for tasks in the database, but made for an confusing, non-transparent extra step when tasks were exported to code. The deletion logic is now entirely gone, for both actual deletions and reversions to code. Improved helptext on the relevant form widget also makes the process slightly more transparent.