Fix some forms to be compatible with PHP 5.3
We have $form passed by reference in a straight FormAPI callback. This fits
with the current form building setup that expects a FormAPI array to be
returned by the callback instead of altering it like ctools did previously.

Note: I've search replaced a lot of these forms and I don't know where to test
them all yet so it may actually break a subform helper that is being called
directly.