Fix array plugins not being loaded on rebuild
Since we can never include plugin include files a second time, we can never get
the plugin array back. We can however make the assumptioin that the file won't
change durring runtime(or not in a way we care to deal with) and use an actual
static to store this. We use a real static because if it is ever whiped, we
can never get the information back and recover.